Explicit casting example

+++ src/com/ajaykmrjha/Main.java
@@ -5,10 +5,9 @@ import java.util.Arrays;
 public class Main {
 
     public static void main(String[] args) {
-//        Implicit casting | when we don't want to lose any data
-//        byte > short > int > long > float > double
+//        Explicit casting 
         double x = 1.1;
-        double y = x + 2;
+        int y = (int) x + 2;
         System.out.println(y);
